requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pdfplumber==0.11.4
python-docx==1.1.2
